added all the parsing made the routines and all the creating functions it works smoothly, just need to add free function and treat the times eaten parameter

+++ src/utils.c
@@ -6,17 +6,17 @@
 /*   By: hkonstan <hkonstan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/19 20:50:31 by hkonstan          #+#    #+#             */
-/*   Updated: 2026/02/19 20:50:55 by hkonstan         ###   ########.fr       */
+/*   Updated: 2026/02/20 15:35:59 by hkonstan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philosophers.h"
 
-long	get_time(long start_time)
+long long	get_time(long start_time)
 {
 	struct timeval	t;
-	long			c_time;
-	long			sim_time;
+	long long		c_time;
+	long long		sim_time;
 
 	gettimeofday(&t, NULL);
 	c_time = t.tv_sec * 1000 + t.tv_usec / 1000;
